backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
mysql-connector-python==8.4.0
python-dotenv==1.0.1
slowapi==0.1.9